# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.ErrorResp:
    properties:
      error:
        type: string
    type: object
...
  services.GetProfileResp:
    properties:
      bookmarks:
        items:
          $ref: '#/definitions/services.ProfilePost'
        type: array
      following:
        items:
          $ref: '#/definitions/services.ProfileUser'
        type: array
      liked_posts:
        items:
          $ref: '#/definitions/services.ProfilePost'
        type: array
    type: object
...
  services.ProfilePost:
    properties:
      id:
        type: string
      title:
        type: string
    type: object
  services.ProfileUser:
    properties:
      id:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the blog0 API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Blog0 API
  version: "1.0"
paths:
...
  /api/v1/me/profile:
    get:
      consumes:
      - application/json
      description: Get user's following, bookmarks, and liked posts (requires authentication)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.GetProfileResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResp'
      security:
      - BearerAuth: []
      summary: Get user profile
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"